🧪 RW8CJF task: isolate command session trace test
