+++ packages/agentplane/src/cli/run-cli.core.test.ts
@@ -222,33 +222,6 @@ describe("runCli", () => {
     }
   });
 
-  it("traces only the declared output node for a metadata-only command", async () => {
-    const root = await mkTempDir();
-    const outPath = path.join(root, "cli-reference.mdx");
-    const previousTrace = process.env.AGENTPLANE_TRACE;
-    process.env.AGENTPLANE_TRACE = "1";
-    const io = captureStdIO();
-    try {
-      const code = await runCli(["docs", "cli", "--out", outPath]);
-      expect(code).toBe(0);
-      expect(io.stderr).toContain('"component":"command-session"');
-      expect(io.stderr).toContain('"event":"preparation_node"');
-      expect(io.stderr).toContain('"capability":"output"');
-      expect(io.stderr).toContain('"node":"output"');
-      expect(io.stderr).not.toContain('"node":"project"');
-      expect(io.stderr).not.toContain('"node":"config"');
-      expect(io.stderr).not.toContain('"node":"command_context"');
-    } finally {
-      if (previousTrace === undefined) {
-        delete process.env.AGENTPLANE_TRACE;
-      } else {
-        process.env.AGENTPLANE_TRACE = previousTrace;
-      }
-      io.restore();
-      await rm(root, { force: true, recursive: true });
-    }
-  });
-
   it("does not load .env for ide sync when dispatch only needs project metadata", async () => {
     const root = await mkGitRepoRoot();
     await writeDefaultConfig(root);
